requirements.txt
Django == 3.2.4
celery == 5.2.6
redis == 4.2.1
django-multiselectfield == 0.1.12
psycopg2 == 2.9.3
Pillow == 7.0.0
django-jsx == 0.5.0